+++ src/v1/sources/shopify/webhookTransformations/serverSideUtlis.js
@@ -5,6 +5,7 @@ const { extractEmailFromPayload } = require('../../../../v0/sources/shopify/util
 const { constructPayload } = require('../../../../v0/util');
 const { INTEGERATION, lineItemsMappingJSON, productMappingJSON } = require('../config');
 const { RedisDB } = require('../../../../util/redis/redisConnector');
+const logger = require('../../../../logger');
 
 /**
  * Returns an array of products from the lineItems array received from the webhook event
@@ -70,7 +71,7 @@ const getCartToken = (event) => event?.cart_token || null;
  * @param {Object} message
  * @param {Object} event
  */
-const handleAnonymousId = async (message, event) => {
+const setAnonymousId = async (message, event) => {
   const anonymousId = getAnonymousIdFromAttributes(event);
   if (isDefinedAndNotNull(anonymousId)) {
     message.anonymousId = anonymousId;
@@ -82,15 +83,22 @@ const handleAnonymousId = async (message, event) => {
       if (redisData?.anonymousId) {
         message.anonymousId = redisData.anonymousId;
       }
+    } else {
+      // have log in case cart_token is not present in the event, attach the log to the writekey
+      logger.error('Cart token not found in the event', {
+        event,
+        sourceId: message.sourceId,
+        workspaceId: message.workspaceId,
+      });
     }
   }
 };
 
 /**
   Handles userId, email and contextual properties enrichment for the message payload
- * @param {Object} message
- * @param {Object} event
- * @param {String} shopifyTopic
+ * @param {Object} message rudderstack message object
+ * @param {Object} event raw shopify event payload
+ * @param {String} shopifyTopic shopify event topic
 */
 const handleCommonProperties = (message, event, shopifyTopic) => {
   if (message.userId) {
@@ -125,6 +133,6 @@ module.exports = {
   getProductsFromLineItems,
   getAnonymousIdFromAttributes,
   getCartToken,
-  handleAnonymousId,
+  setAnonymousId,
   handleCommonProperties,
 };
